Avoid empty data

+++ ServerPHPScripts/index.php
@@ -2,6 +2,10 @@
 include "db.php";
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $data = json_decode(file_get_contents('php://input'), true);
+    if (empty($data['objects'])) {
+        echo "Failed to read save data";
+        exit();
+    }
     mysqli_query($GLOBALS['mysqli'], "DELETE FROM objects WHERE 1");
     foreach ($data['objects'] as $object) {
         mysqli_query($GLOBALS['mysqli'], "INSERT INTO objects (prefabId, contentType, content, posX, posY, posZ, rotX, rotY, rotZ, scaleX, scaleY, scaleZ)");
